Move action creators, reducers and selectors from common to src folder.
