+++ services/core-web/common/reducers/mineReducer.js
@@ -17,7 +17,6 @@ const initialState = {
   mineDocuments: [],
   subscribedMines: [],
   mineComments: [],
-  mineAlerts: [],
 };
 
 export const mineReducer = (state = initialState, action) => {
@@ -70,11 +69,6 @@ export const mineReducer = (state = initialState, action) => {
         ...state,
         mineComments: action.payload.records,
       };
-    case actionTypes.STORE_MINE_ALERTS:
-      return {
-        ...state,
-        mineAlerts: action.payload.records,
-      };
     default:
       return state;
   }
@@ -96,6 +90,5 @@ export const getCurrentUserVerifiedMines = (state) => state[MINES].currentUserVe
 export const getCurrentUserUnverifiedMines = (state) =>
   state[MINES].currentUserUnverifiedMinesMines;
 export const getMineComments = (state) => state[MINES].mineComments;
-export const getMineAlerts = (state) => state[MINES].mineAlerts;
 
 export default mineReducerObject;
