+++ services/core-web/common/selectors/mineSelectors.js
@@ -10,7 +10,6 @@ export const getMineBasicInfoList = (state) => mineReducer.getMineBasicInfoList(
 export const getMineDocuments = (state) => mineReducer.getMineDocuments(state);
 export const getSubscribedMines = (state) => mineReducer.getSubscribedMines(state);
 export const getMineComments = (state) => mineReducer.getMineComments(state);
-export const getMineAlerts = (state) => mineReducer.getMineAlerts(state);
 
 export const getIsUserSubscribed = createSelector(
   [getSubscribedMines, getMineGuid],
