+++ services/core-web/src/constants/actionTypes.js
@@ -1,2 +1,6 @@
 export const DOCUMENTS = "DOCUMENTS";
 export const STORE_DOCUMENT_CONTEXT_TEMPLATE = "STORE_DOCUMENT_CONTEXT_TEMPLATE";
+
+// Mine Alerts
+export const STORE_MINE_ALERTS = "STORE_MINE_ALERTS"
+export const CLEAR_MINE_ALERT = "CLEAR_MINE_ALERT";
