+++ services/core-web/src/constants/reducerTypes.js
@@ -9,3 +9,10 @@ export const EXPORT_NOTICE_OF_WORK_APPLICATION_DOCUMENT =
 export const GENERATE_EXPLOSIVES_PERMIT_DOCUMENT = "GENERATE_EXPLOSIVES_PERMIT_DOCUMENT";
 export const GET_EXPLOSIVES_PERMIT_DOCUMENT_CONTEXT_TEMPLATE =
   "GET_NOTICE_OF_WORK_APPLICATION_DOCUMENT_CONTEXT_TEMPLATE";
+
+// Alerts
+export const MINE_ALERTS = "MINE_ALERTS";
+export const CREATE_MINE_ALERTS = "CREATE_MINE_ALERTS";
+export const UPDATE_MINE_ALERT = "UPDATE_MINE_ALERT";
+export const DELETE_MINE_ALERT = "DELETE_MINE_ALERT";
+export const GET_MINE_ALERTS = "GET_MINE_ALERTS";
