+++ services/core-web/src/reducers/rootReducer.js
@@ -55,6 +55,7 @@ import {
 } from "@common/reducers";
 import * as reducerTypes from "@common/constants/reducerTypes";
 import { createReducer } from "@common/utils/helpers";
+import mineAlertReducer  from "./mineAlertReducer";
 import networkReducer from "./networkReducer";
 import documentReducer from "./documentReducer";
 
@@ -87,6 +88,7 @@ export const reducerObject = {
   ...activityReducer,
   ...tailingsReducer,
   ...damReducer,
+  ...mineAlertReducer,
   [reducerTypes.CREATE_PARTY]: createReducer(networkReducer, reducerTypes.CREATE_PARTY),
   [reducerTypes.UPDATE_PARTY]: createReducer(networkReducer, reducerTypes.UPDATE_PARTY),
   [reducerTypes.GET_PARTIES]: createReducer(networkReducer, reducerTypes.GET_PARTIES),
