+++ services/core-web/src/selectors/mineAlertSelectors.js
@@ -1,0 +1,4 @@
+import * as mineAlertReducer from "@/reducers/mineAlertReducer";
+
+// eslint-disable-next-line import/prefer-default-export
+export const getMineAlerts = (state) => mineAlertReducer.getMineAlerts(state);
