+++ services/minespace-web/common/actions/mineActions.js
@@ -45,12 +45,3 @@ export const storeEpicInfo = (payload) => ({
   type: ActionTypes.STORE_MINE_EPIC_INFO,
   payload,
 });
-
-export const storeMineAlerts = (payload) => ({
-  type: ActionTypes.STORE_MINE_ALERTS,
-  payload,
-});
-
-export const clearMineAlert = () => ({
-  type: ActionTypes.CLEAR_MINE_ALERT,
-});
