+++ services/minespace-web/common/constants/actionTypes.js
@@ -187,7 +187,3 @@ export const CLEAR_TAILINGS_STORAGE_FACILITY = "CLEAR_TAILINGS_STORAGE_FACILITY"
 
 // Dams
 export const STORE_DAM = "STORE_DAM";
-
-// Mine Alerts
-export const STORE_MINE_ALERTS = "STORE_MINE_ALERTS"
-export const CLEAR_MINE_ALERT = "CLEAR_MINE_ALERT";
