+++ services/minespace-web/common/constants/reducerTypes.js
@@ -298,9 +298,3 @@ export const DAMS = "DAMS";
 export const CREATE_DAM = "CREATE_DAM";
 export const UPDATE_DAM = "UPDATE_DAM";
 export const GET_DAM = "GET_DAM";
-
-// Alerts
-export const CREATE_MINE_ALERTS = "CREATE_MINE_ALERTS";
-export const UPDATE_MINE_ALERT = "UPDATE_MINE_ALERT";
-export const DELETE_MINE_ALERT = "DELETE_MINE_ALERT";
-export const GET_MINE_ALERTS = "GET_MINE_ALERTS";
